Add support to select memfault specific components
Due to PlatformIO nature and our current definition of the library.json
all the components from the SDK are built by default.

This lead to some wasted space and enabled feature the user may not use.

Add a build script so components that should be activated can be passed
as a whitespace separated list within the `MEMFAULT_COMPONENTS` macro.
The macro defaults to `core util panics metrics`.

Signed-off-by: Kenfe-Mickael Laventure <mickael@impulselabs.com>